# docs/api-specs/temporal.openapi.json
{
  "components": {
    "schemas": {
...
      "Device": {
        "description": "Device data for dropdown population.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "Device",
        "type": "object"
      },
...
      "DeviceInterface": {
        "description": "Device interface data for dropdown population.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "DeviceInterface",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "Bearer JWT used by CLI and machine clients on svc-* endpoints. Authentication is required by default, but deployments may disable it with [auth] required = false. Browser OIDC sessions, mTLS, SPIFFE identities, and trusted gateway identity headers may also satisfy authentication outside this generated client flow.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Temporal Workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/parameter/device": {
      "get": {
        "description": "Return a list of filtered devices.",
        "operationId": "get_devices_v1_parameter_device_get",
        "parameters": [
          {
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "tenant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant"
            }
          },
          {
            "in": "query",
            "name": "device_type_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Type Id"
            }
          },
          {
            "in": "query",
            "name": "manufacturer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manufacturer"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "description": "Limit to NVIDIA Config Manager-managed devices",
            "in": "query",
            "name": "managed_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "Limit to NVIDIA Config Manager-managed devices",
              "title": "Managed Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  },
                  "title": "Response Get Devices V1 Parameter Device Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Devices",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/device-id": {
      "get": {
        "description": "Convert device hostname to device ID.\n\nArgs:\n    device_name: The hostname/name of the device.\n\nReturns:\n    Device object with id and name.\n\nRaises:\n    HTTPException: If device not found or multiple devices match.",
        "operationId": "get_device_id_by_name_v1_parameter_device_id_get",
        "parameters": [
          {
            "in": "query",
            "name": "device_name",
            "required": true,
            "schema": {
              "title": "Device Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Device ID by Name",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/device/{device_id}/interfaces": {
      "get": {
        "description": "Return the interfaces belonging to a device.",
        "operationId": "get_device_interfaces_v1_parameter_device__device_id__interfaces_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceInterface"
                  },
                  "title": "Response Get Device Interfaces V1 Parameter Device  Device Id  Interfaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Device Interfaces",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/device/{device_id}/password_users": {
      "get": {
        "description": "Get available password users and their secret names for a device.\n\nArgs:\n    device_id: The UUID of the device.\n\nReturns:\n    List of password users with their provider-normalized secret names.",
        "operationId": "get_device_password_users_v1_parameter_device__device_id__password_users_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Secret"
                  },
                  "title": "Response Get Device Password Users V1 Parameter Device  Device Id  Password Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Device Password Users",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/device/{device_id}/secret_types": {
      "get": {
        "description": "Get available secret types from the configured DCIM provider.\n\nArgs: device_id: The UUID of the device.\n\nReturns:\n    List of secret types plus versions from config context.",
        "operationId": "get_device_users_with_versions_v1_parameter_device__device_id__secret_types_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Device Users With Versions V1 Parameter Device  Device Id  Secret Types Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Secret Types for Device",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/device/{device_id}/secrets": {
      "get": {
        "description": "Return a list of secrets available for a device.\n\nArgs:\n    device_id: The UUID of the device to fetch secrets from\n\nReturns:\n    List of secrets returned by the configured DCIM provider",
        "operationId": "get_device_secrets_v1_parameter_device__device_id__secrets_get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Secret"
                  },
                  "title": "Response Get Device Secrets V1 Parameter Device  Device Id  Secrets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Device Secrets",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/diagnostics/commands": {
      "get": {
        "description": "Return the diagnostics command catalog for the requested platform(s).\n\nPass one or more ``?platform=<slug>`` query params (e.g. ``cumulus-linux``).\nWhen no platform is given, commands from all platforms are returned.\nCommands that appear on multiple platforms are deduplicated by name.",
        "operationId": "get_diagnostics_commands_v1_parameter_diagnostics_commands_get",
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommandEntry"
                  },
                  "title": "Response Get Diagnostics Commands V1 Parameter Diagnostics Commands Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Diagnostics Commands",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/location": {
      "get": {
        "description": "Return a list of NVIDIA Config Manager-managed sites.",
        "operationId": "get_locations_v1_parameter_location_get",
        "parameters": [
          {
            "in": "query",
            "name": "location_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  },
                  "title": "Response Get Locations V1 Parameter Location Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Locations",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/namespace-tag": {
      "get": {
        "description": "Return the configured DCIM provider's namespace tag choices.",
        "operationId": "get_namespace_tags_v1_parameter_namespace_tag_get",
        "parameters": [
          {
            "description": "Limit to namespace tags at this location",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limit to namespace tags at this location",
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  },
                  "title": "Response Get Namespace Tags V1 Parameter Namespace Tag Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Namespace Tags",
        "tags": [
          "parameters"
        ]
      }
    },
    "/v1/parameter/overlay": {
      "get": {
        "description": "Return overlays, optionally filtered by location and isolation type.",
        "operationId": "get_overlays_v1_parameter_overlay_get",
        "parameters": [
          {
            "description": "Limit to overlays at this location",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limit to overlays at this location",
              "title": "Location"
            }
          },
          {
            "description": "Limit to overlays with this isolation type",
            "in": "query",
            "name": "isolation_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limit to overlays with this isolation type",
              "title": "Isolation Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Overlay"
                  },
                  "title": "Response Get Overlays V1 Parameter Overlay Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Overlays",
        "tags": [
          "parameters"
        ]
      }
    },
...
  }
}